pyproject.toml
[tool.poetry]
name = "hikari-template"
version = "0.1.2"
description = "A Hikari template meant to quickly setup a Discord bot using Hikari."
authors = ["Enrique Bos <enriquebos7090@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
hikari = {extras = ["speedups"], version = "^2.0.0.dev120"}
hikari-lightbulb = "^2.3.3"
uvloop = "^0.17.0"
pydantic = "^2.1.1"
pydantic-settings = "^2.0.2"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
ruff = "^0.0.280"
mypy = "^1.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.10"
strict = true
warn_unreachable = true
implicit_reexport = true

# Disable specific strict options since we don't own all the modules
ignore_missing_imports = true
disallow_untyped_decorators = false

# Enable plugins for specfic packages
plugins = ["pydantic.mypy"] # "sqlalchemy.ext.mypy.plugin"

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[[tool.mypy.overrides]]
module = 'lightbulb.*'

[tool.ruff]
target-version = "py311"

# Naming rules -> https://beta.ruff.rs/docs/rules/
select = [
    "F", "E", "W", "C90", "I",
    "N", "D", "UP", "ANN", "S",
    "BLE", "FBT", "B", "A", "COM",
    "C4", "DTZ", "EM", "ISC", "ICN",
    "G", "PIE", "T20", "PYI", "PT",
    "PT", "Q", "RSE", "RET", "SLF",
    "SIM", "TID", "TCH", "INT", "ARG",
    "PTH", "ERA", "PD", "PGH", "PL",
    "PLC", "PLE", "PLR", "PLW", "TRY",
    "NPY", "RUF"]

# Ignoring
# `no-blank-line-before-class` (D211) in favor of `one-blank-line-before-class` (D203)
# `multi-line-summary-first-line` (D212) in favor of `multi-line-summary-second-line` (D213)
# Forced module docstring (D100)
# Missing docstring in public function (D103)
# Missing docstring in public package (D104)
ignore = ["D211", "D212", "D100", "D103", "D104"]
